apply fraud ml service latest

+++ claim-service/src/main/java/group102/insurancefraud/dto/response/ShapResultDTO.java
@@ -22,7 +22,7 @@ public class ShapResultDTO {
     private String summary;
 
     @JsonProperty("confidence")
-    private Integer confidence;
+    private Double confidence;
 
     @JsonProperty("method")
     private String method;
